+++ src/atoms/image/image.tsx
@@ -3,7 +3,6 @@ import { ImageTypes, AmpImgPropTypes } from "./types";
 import { focusedImagePath } from "../../helpers";
 import { withConfig } from "../../context";
 import { LightboxGallery } from "../lightbox-gallery";
-import { Helmet } from "react-helmet";
 
 export const BaseImage = ({
   metadata,
@@ -16,7 +15,6 @@ export const BaseImage = ({
   opts = {},
   config,
   lightbox = true,
-  preloadImage,
   ...rest
 }: ImageTypes) => {
   const cdnImage = config.publisherConfig["cdn-image"];
@@ -52,25 +50,13 @@ export const BaseImage = ({
       value.height = imgAspectRatio[1].toString();
   }
 
-  const imageComponent = lightbox ? (
+  return lightbox ? (
     <Fragment>
       <LightboxGallery />
       <amp-img {...value} lightbox={lightbox} />
     </Fragment>
   ) : (
     <amp-img {...value} />
-  );
-
-  return preloadImage ? (
-    <Fragment>
-      <Helmet>
-        <link rel="preload" as="image" href={srcSmall} media="(max-width: 1000px)" crossorigin="anonymous" />
-        <link rel="preload" as="image" href={srcBig} media="(min-width: 1001px)" crossorigin="anonymous" />
-      </Helmet>
-      {imageComponent}
-    </Fragment>
-  ) : (
-    imageComponent
   );
 };