+++ src/molecules/related-stories/related-stories.tsx
@@ -6,6 +6,9 @@ import { withStoryAndConfig } from "../../context";
 import { base64FallbackImage } from "../../helpers/image-helpers";
 import { Spacer } from "../../atoms";
 import get from "lodash.get";
+import { StoryPageSlots } from "../../molecules/slots";
+
+const { RelatedStoryCardSlot } = StoryPageSlots;
 
 export const Heading = styled.h2`
   font-size: ${(props) => props.theme.font.size.l};
@@ -26,17 +29,35 @@ export const RelatedStoriesBase = ({
       <Spacer token="m" />
       <Heading>{heading}</Heading>
       <div>
-        {relatedStories.map((relatedStory) => (
-          <RelatedStoryCard
-            key={relatedStory.id}
-            story={relatedStory}
-            fallbackSrc={get(config, ["ampConfig", "fallback-image-url"], base64FallbackImage)}
-            aspectRatio={aspectRatio}
-          />
+        {relatedStories.map((relatedStory, idx) => (
+          <Fragment key={relatedStory.id}>
+            <RelatedStoryCard
+              story={relatedStory}
+              fallbackSrc={get(config, ["ampConfig", "fallback-image-url"], base64FallbackImage)}
+              aspectRatio={aspectRatio}
+            />
+            <RelatedStoryCardSlot story={story} config={config} index={idx} relatedStory={relatedStory} />
+          </Fragment>
         ))}
       </div>
     </Fragment>
   );
 };
+
+/**
+ * RelatedStories Component uses the RelatedStoryCard molecule component internally.
+ *
+ * ```js
+ * <RelatedStories stories={relatedStories} />
+ * ```
+ *
+ * @param {Object} props Object containing parameters passed to the render prop
+ * @param {Object} props.stories Required. It takes in related stories object.
+ * @param {Object} props.config config object
+ * @param {Object} props.heading Optional. A String which serves as a title for this collection. If not passed, defaults to "Also Read" text.
+ * @param {Object} props.aspectRatio Optional. An Array of width and height. If not passed, defaults to [16,9].
+ *  @category Molecules
+ * @component
+ */
 
 export const RelatedStories = withStoryAndConfig(RelatedStoriesBase);