+++ src/molecules/related-story-card/related-story-card.stories.tsx
@@ -6,7 +6,7 @@ import { relatedStories } from "../../__fixtures__/related-stories";
 import { config, textStory } from "../../__fixtures__";
 
 const sampleRelatedStory = relatedStories[0];
-const relatedStoryEithoutHeroImage = relatedStories[1];
+const relatedStoryWithoutHeroImage = relatedStories[1];
 
 storiesOf("Related Story Card", module)
   .addDecorator((story) => (
@@ -18,10 +18,10 @@ storiesOf("Related Story Card", module)
   .add("Without HeroImage", () => (
     <RelatedStoryCard
       aspectRatio={[16, 9]}
-      story={relatedStoryEithoutHeroImage}
+      story={relatedStoryWithoutHeroImage}
       fallbackSrc="https://gumlet.assettype.com/quintype-demo%2F2019-02%2Fae4829d6-e230-4d66-9b88-2a4db87a2060%2F64778.jpg?w=640&auto=format%2Ccompress"
     />
   ))
   .add("Without HeroImage & without fallback src", () => (
-    <RelatedStoryCard aspectRatio={[16, 9]} fallbackSrc="" story={relatedStoryEithoutHeroImage} />
+    <RelatedStoryCard aspectRatio={[16, 9]} fallbackSrc="" story={relatedStoryWithoutHeroImage} />
   ));