+++ src/molecules/slots/story-page-slots/default-story-card/index.ts
@@ -1,0 +1,2 @@
+import { DefaultStoryCardSlot } from "./default-story-card-slot";
+export { DefaultStoryCardSlot };