+++ src/molecules/slots/story-page-slots/live-blog-card/live-blog-card-slot.tsx
@@ -1,5 +1,5 @@
-import { withStoryAndConfig } from "../../../../../context";
-import { getSlotFromConfig } from "../../../helpers";
+import { withStoryAndConfig } from "../../../../context";
+import { getSlotFromConfig } from "../../helpers";
 
 const LiveBlogCardSlotBase = ({ story, config, index, card }) => {
   const liveBlogCardRender = getSlotFromConfig({
@@ -13,7 +13,7 @@ const LiveBlogCardSlotBase = ({ story, config, index, card }) => {
 /**
  * live-blog-card is a story page slot. It's placed after every card on a live blog type story.
  *
- * @category Atoms
+ * @category Molecules
  * @module live-blog-card
  * @param {Props} props
  * @param {Story} props.story