+++ src/templates/generic-story/generic-story.test.tsx
@@ -1,9 +1,47 @@
 import React from "react";
-import { shallow } from "enzyme";
+import { shallow, mount } from "enzyme";
 import { GenericStory } from "./generic-story";
 import { InfiniteScroll, Subscription } from "../../atoms";
 import { textStory, config } from "../../__fixtures__";
 import { TopAd, BottomAd } from "../../molecules/ads";
+import { StoryPageSlots } from "../../molecules/slots";
+import cloneDeep from "lodash.clonedeep";
+
+const { TopSlot, BottomSlot, RelatedStoryCardSlot } = StoryPageSlots;
+
+const genDummyCard = () => {
+  const dummyId = String(Math.ceil(Math.random() * 100000000000));
+  return {
+    "story-elements": [
+      {
+        description: "",
+        "page-url": "foo/bar",
+        type: "text",
+        "family-id": "123abc",
+        title: "",
+        id: "def567",
+        metadata: {},
+        subtype: null,
+        text: "<p>lorem ipsum</p>"
+      }
+    ],
+    "card-updated-at": 1581327522163,
+    "content-version-id": "efaf78de-c90b-4d15-b040-c84ebb29cabf",
+    "card-added-at": 1581327522163,
+    status: "draft",
+    id: dummyId,
+    "content-id": dummyId,
+    version: 1,
+    metadata: {
+      "social-share": {
+        shareable: false,
+        title: "#Sponsored: Empowering India’s Future Lawyers and Leaders",
+        message: null,
+        image: null
+      }
+    }
+  };
+};
 
 describe("GenericStory Template", () => {
   it("should render", () => {
@@ -39,5 +77,17 @@ describe("GenericStory Template", () => {
     expect(wrapper.find(TopAd).prop("templateName")).toBe("default");
     // expect(wrapper.find(BodyAd).prop("templateName")).toBe("default");
     expect(wrapper.find(BottomAd).prop("templateName")).toBe("default");
+  });
+  it("should render all generic story page slots", () => {
+    const dummyStory = cloneDeep(textStory);
+    dummyStory.cards = [];
+    for (let i = 0; i < 5; i++) {
+      dummyStory.cards.push(genDummyCard());
+    }
+
+    const wrapper = mount(<GenericStory story={dummyStory} config={config} />);
+    expect(wrapper.find(TopSlot).length).toBe(1);
+    expect(wrapper.find(BottomSlot).length).toBe(1);
+    expect(wrapper.find(RelatedStoryCardSlot).length).toBe(4);
   });
 });